// packages/frontend/src/types/schemas/prisma-generated/inputTypeSchemas/EmailABTestResultUncheckedCreateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';

export const EmailABTestResultUncheckedCreateInputSchema: z.ZodType<Prisma.EmailABTestResultUncheckedCreateInput> = z.object({
  id: z.string().uuid().optional(),
  testId: z.string(),
  variant: z.string(),
  emailId: z.string(),
  opened: z.boolean().optional(),
  clicked: z.boolean().optional(),
  createdAt: z.coerce.date().optional()
}).strict();

export default EmailABTestResultUncheckedCreateInputSchema;